Allow `concat` in SpirV macro
The `include_spirv` macros was set to accept literals only.
Unfortunately, this means that other macros, such as `concat` and `env`
aren't accepted, making it unusable with build script-compiled shaders.
This change allows those macros to be used to load the SpirV at the
expense of a potentially less helpful error message as the `literal`
requirement is differed to the `include_bytes` invocation.