Add a trainilg ']' to the trace if player fails to load the RON file
In case of a panic, the RON trace ends without appending a ']' at
the end. Loading this trace as it is will cause a failure while
loading as this doesn't follow the specifications for a valid
RON file.
In case of an error while loading the RON file, we append the ']'
at the end of file and retry. If the error was indeed a missing ']'
the trace is loaded correctly.
If the load fails second time, this means that the trace needs a
manual review for correctness but appending ']' solves most of the
cases of invalid RON file obtained as a result of panic.