[wgsl-in]: Report unknown identifiers as Error::UnknownIdent.
Move the parsing of type construction expressions into a function of its own,
which can return `None` when the identifier at hand is not the name of a type.
This means that `parse_primary_expression` doesn't have to commit to
interpreting the expression as a type construction too early, allowing it to
report unknown identifiers as Error::UnknownIdent.