Move validation from `Device::create_texture_from_hal()` to `Device::create_texture()`
---

Users encountered a segfault upon creating a texture when mip_level_count was set to 0 with (at least) Vulkan.
This is due to wgpu's validation taking place after the API's texture is created.

This change moves wgpu's validation to before the API's texture is created, introducing a friendly error instead of a segfault.