Fix wasm examples failing to compile
They were broken by 476b6a1, because apparently enabling a feature in any package enables it for the whole workspace, and the wasm examples fail to compile with vulkan enabled.

It wasn't failing in CI because CI specifies `-p wgpu`, which seems to stop `wgpu-info`'s enabling of features having an effect.