Add `Buffer::size()` and `Buffer::usage()`.
https://gpuweb.github.io/gpuweb/#buffer-interface

Implementation note: it would save 8 bytes of memory to ask the backend
for the size, instead of storing it separately, since many backends have
the size stored anyway. But, that seemed like it would overcomplicate
the architecture for tiny memory savings, and `usage` is already stored
in the `wgpu::Buffer` anyway (though only used once, prior to now).