vulkan: use effective api version for determining device features
wgpu wrongly assumes that the instance version will always be the maximum value. This assumption can be broken by using Instance::from_raw. This pull request fixes that issue by calculating the "effective" api version the device supports, which is the lower of the instance version and physical device version. More information can be found in the Vulkan specification's documentation for VkPhysicalDeviceProperties.