Allow creating external texture views and samplers
As a part of the work to make external memory function, there needs to be a way to create image views and samplers externally. On Android and Linux, with the VK_EXT_image_drm_modifier and VK_ANDROID_external_memory_android_hardware_buffer extensions, some formats may require a Ycbcr sampler as provided by the VK_KHR_sampler_ycbcr_conversion extension. This sampler type needs to be provided both when creating a Sampler and the TextureViews for which the sampler is used in.

Like any of the prior pull requests, the best way I've seen so far is to expose the logic to allow external mechanisms to create the objects with special extensions outside of wgpu-hal and provide a way to import the handles into wgpu-hal.