Introduce boxes and vectors to reduce Result sizes.
The following error variants were large enough to trip Clippy's
`result_large_err` lint:

- RenderPassCompatibilityError::IncompatibleColorAttachment
- CreateShaderModuleError::Validation

Large error types are a problem, because they make the `Result` type
large, requiring the program to copy around large numbers of bytes
even in the success case. Since the failure case is rare, it's usually
not a problem to just heap-allocate the contents to keep the `Ok` case
small.