Allow unspecified bits in the API's bitflags
The bitflags_serde_shim treats unspecified bits as errors during deserialization. To make the browser integration easier, we want the unsepcified bits (which can be produced by web content) to turn into validation errors instead of panics during deserialization.

Thankfully the bitflags_serde_shim crate is quite trivial, so this PR inlines the macro with the required modification to not check the bits. In addition, bitflags_serde_shim feature is removed, the Serialize and Deserialize implementations are aways the same and only their presence depend on the trace and replay feature flags.