Fix surface creation cfgs in example framework.
Expand CI coverage so that we try building `wgpu/examples/framework.rs` on
Emscripten.

The `framework.rs` library used to check:

    #[cfg(not(target_arch = "wasm32"))]

to decide whether to call `instance.create_surface` or attempt to
support creating a surface from an offscreen canvas. This is not
correct: `emscripten` should be be treated as a non-web target; the
instance doesn't have a `create_surface_from_offscreen_canvas` method.

Instead, `framework.rs` should check:

    #[cfg(any(not(target_arch = "wasm32"), target_os = "emscripten"))]

to treat the Emscripten target like a native target.