Make RequestAdapterOptions.power_preference optional
The WebGPU docs state that the `powerPreference` field is optional -
it can be set to `undefined`, which provides no hint.

Currently, wgpu only allows using an explicit LowPower or HighPerformance
hint. This can cause a secondary backend to be chosen over a primary
backend when the secondary backend is the only one matching the power
preference - with no way to ensure that a primary backend takes
priority.

This commit changes `RequestAdapterOptions.power_preference` to
store an `Option<PowerPreference>`, and removes the `Default`
impl from `PowerPreference`. Under WebGPU, we leave
`GPURequestAdapterOptions.powerPreference` when our value
is `None`. Under native, we skip the backend preference logic.