Add details to `InstanceError` and `CreateSurfaceError`.
This will allow developers, particularly those in communication with
users with hardware they don't own, to diagnose “why won't this work
on my machine” better, without needing to obtain logs in addition to
the error value.

Note that this is a semver-breaking change because `InstanceError` and
`CreateSurfaceError` no longer implement `PartialEq`. This could be
avoided by pointer comparison of the internal `Arc`, but I think it is
more typical Rust to not implement `PartialEq` for error types unless
there is a strong motivation to and their contents are tightly
controlled.