Give test image snapshots distinctive names.
Include the adapter name and driver name in the filenames generated to
store the `-actual.png` and `-difference.png` files holding
screenshots of test program output.

Previously, the filename only incorporated the backend, but since we
test the Vulkan backend on both Mac and Windows, one was overwriting
the other in the artifacts.