Implement device validity, and make it invalid when device is lost.
This tracks device validity explicitly, and short-circuits almost all
operations on devices when they are invalid. This doesn't replace the
storage of the device with an Error, because it's important for
poll_devices to be able to monitor the devices (even if invalid) for
emptiness so they can be cleaned up.