wgpu-hal: vulkan: ensure entries to PhysicalDeviceCapabilities use effective api version
This was the cause of an error while testing wgpu-hal on agxv. Specifically I was getting an error about running out of memory while creating the "zero init buffer".
This was occuring because PhysicalDeviceCapabilities was reading the instance api version and not the device version. But furthermore this is wrong because the actual api version that can be used is the lower of the instance's api version and device's api version.

agxv does not implement maintenance3 yet, so what was being returned for a max memory allocation size was 0, because ash initialized the value to 0. Then when the gpu allocator is created with a maximum allocation size of 0, all gpu memory allocations fail.

Signed-off-by: i509VCB <git@i509.me>