Add a LoseDeviceClosure to wgpu-core::Device, route it to wgpu.
The LoseDeviceClosure is called in two cases:

1) It is returned as one of the closures from `maintain` if
`Device.destroy` has been called. This helps implement the spec for
`destroy` that says that "lose the device" is the final step after the
queues have emptied.
2) It is called directly by `Device.lose`.

To accomplish this, `Device.valid` is set to false in `destroy`, which
is not compliant with the spec but is necessary to ensure that no new
work is enqueued after `destroy`. This means that such calls will return
`DeviceError::InvalidDevice` Errs during this period, similarly to how
they already do after "lose the device". Possibly calls during one or
both of these periods should instant fail silently and return `Ok`, to
be handled in a future patch.