Add a test of device_set_lose_device_closure.
In order to supply a DeviceLoseClosure, the test harness has to be given
access to wgpu-core, so this patch modifies the Cargo files to do that.

The new test confirms that the closure gets the expected value *if it is
run*. As noted in a comment in the test, a more robust test that
confirms the closure has been called requires the test function to
become async, which imposes restrictions on how the test is run.