[wgsl-in] Remove unused expressions from Module::const_expressions.
After building a Naga module from the WGSL AST, make a pass over the
module to determine which expressions in `Module::const_expressions`
are actually used, and compact out all unused expressions from that
arena, adjusting references appropriately.

This isn't necessary on trunk, but it will be when we add abstract
types to the WGSL front end. The evaluation of expressions containing
abstract values will naturally introduce lots of 64-bit float and
integer literals to the constant arena, which the validator would
rightly reject. However, because all abstract values are converted to
concrete types before appearing as part of the module's final code,
all such 64-bit expressions should be unused.