Prefer `Arc::into_inner` over `Arc::try_unwrap`.
As explained in its Rust documentation, calling `Arc::try_unwrap(a)`
where the intention is to drop `a` if it is not uniquely owned suffers
from a race condition: if two threads call `try_unwrap` simulaneously
on the last two `Arc`s referring to a value, they will both see a
reference count greater than one, and then both drop the `Err`, with
neither thread getting the value. Because `try_unwrap` returns the
`Arc` in its `Err`, there is a check-versus-use gap. `Arc::into_inner`
returns an `Option` of the result, and guarantees that checking the
count and dropping the `Arc` are indivisible.